==> books/3 model.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import LinearSVC
from sklearn.linear_model import SGDClassifier
from sklearn import metrics
from sklearn.dummy import DummyClassifier

load data

lang = 'en'
target = 'sentiment'
ds = pd.read_csv('../dataset/wiki/opinions_annotated.csv')
ds = ds[ds.lang==lang]

build train and test dataset for classification

stance_dist=ds.groupby(target).size()
stance_dist

since the labels are balanced we take all of them

sample = ds[~ds[target].isnull()]
sample.shape

X = sample[['text', 'type']]
y = sample[target]

split in train and test dataset

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

print(f'train size: {X_train.shape} , test size: {X_test.shape}')

extract text features

from sklearn.feature_extraction.text import TfidfVectorizer
vectorizer = TfidfVectorizer(stop_words='english', strip_accents='ascii')
X_train_vec = vectorizer.fit_transform(X_train.text)
X_test_vec = vectorizer.transform(X_test.text)
print(vectorizer)
print(f'text features: {X_train_vec.shape[1]}')

build and evaluate classifiers

acc_scores = {}
X_test = X_test.copy()
X_test['y_test'] = y_test

def benchmark_model(name):
    print(clf.fit(X_train_vec,y_train))
    y_pred = clf.predict(X_test_vec)
    X_test.loc[:,name] = y_pred
    score = metrics.accuracy_score(y_test, y_pred)
    f1_micro = metrics.f1_score(y_test, y_pred, average='micro')
    f1_macro = metrics.f1_score(y_test, y_pred, average='macro')
    acc_scores[name]=(name, score, f1_micro, f1_macro)
    print(f'score: {score}')

clf = DummyClassifier(strategy='stratified')
benchmark_model('random')

clf = MultinomialNB()
benchmark_model('NB')

build SGD classifier

clf= SGDClassifier()
benchmark_model('SGD')

clf = LinearSVC(class_weight='balanced')
benchmark_model('LinearSVC')

feature_names = vectorizer.get_feature_names()
if feature_names:
    feature_names = np.asarray(feature_names)

keywords=[]
labels = ['negative', 'neutral', 'positive']
print("top 10 keywords per class:")
for i, label in enumerate(labels):
    top10 = np.argsort(clf.coef_[i])[-10:]
    #print("%s: %s" % (label, " ".join(feature_names[top10])))
    keywords.append((label, " ".join(feature_names[top10])))
    
keywords

ds_keywords = pd.DataFrame(keywords, columns=['label', 'top 10 words'])
ds_keywords.to_csv(f'../results/keywords_{target}_{lang}.csv', index=False)
ds_keywords.head()

acc_ds = pd.DataFrame(list(acc_scores.values()), columns=['model', 'accuracy', 'f1 micro', 'f1 macro'])
acc_ds.to_csv(f'../results/opinions_f1{target}_{lang}.csv', index=False)
acc_ds.head()

types = X_test.groupby(['type'])


for name, group in types:
    for clf_name in list(acc_scores.keys()):
        #TODO: use only pos and neg inside groups
        fscore=metrics.f1_score(group['y_test'], group[clf_name], average='micro') 
        X_test.loc[X_test.type==name, 'f1' + clf_name] = fscore
        #print(fscore)

f1_types = X_test.groupby('type').mean()
f1_types = f1_types.reset_index()
f1_types = f1_types[['type','f1random', 'f1NB','f1SGD','f1LinearSVC']]
f1_types.to_csv(f'../results/opinions_f1{target}_types_{lang}.csv', index=False)
f1_types.head()

import os

out_file = '../dataset/wiki/opinions_predicted.csv'
ds_out = None

if os.path.exists(out_file):
    print('reading output file')
    ds_out = pd.read_csv(out_file)
else:
    print('reading annotated file')
    ds_out = pd.read_csv('../dataset/wiki/opinions_annotated.csv')
    

ix = ds_out.lang==lang
X_ds = vectorizer.transform(ds_out[ix].text)
ds_out.loc[ix, target + '_pred'] = clf.predict(X_ds)
ds_out.to_csv(out_file, index=False)
print(f'predicting target:{target}, lang:{lang} size:{ds_out.shape}.. done')

ds_out.columns

print(ds_out[ds_out[target + '_pred'].isnull()].shape)
